리뷰조회, 등록, 마이페이지 조회, 정보 수정 api

+++ back/src/main/java/teamparkinglot/parkinggo/review/ReviewPostDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReviewPostDto {
-    private Long star;
+    private Double star;
     private String body;
 }
